start centralizing sync/async exn classification
For now, isolate the code that tries to
catch-all-exceptions-except-asynchronous-ones into common combinators, to have
it in one place.

The problem with this idea is that mere classification of exceptions is not an
indicator of how they were thrown, therefore such logic is not entirely correct.
Code that relies on that should move to expecting specific exception types
instead. See, for example,
http://www.haskell.org/pipermail/haskell-cafe/2013-July/107694.html.

Still, a workable band-aid to make it compile.